Fix dereferencing bug
Parser function returns a pointer. However, this pointer is a local
variable. When the function exits, the local variable is destroyed; the
function's return value is not valid. This creates a bug if the
function's return value is subsequently dereferenced.

Fix - allocate memory manually. Now, the memory stays good until freed,
i.e. the function's return value is valid until `free()` is called. This
call is moved out of the function and into the main program (the caller
of the function has to free the memory, not the function itself).